requirements.txt
Reframe-HPC>=4.9.0
pyfirecrest==2.5.0
python-hostlist==1.23.0
